requirements.txt
aiosqlite==0.21.0
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
apprise==1.9.5
asgi-lifespan==2.1.0
asyncpg==0.31.0
attrs==25.4.0
beartype==0.22.7
boto3==1.42.1
botocore==1.42.1
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
coolname==2.2.0
cryptography==46.0.3
dateparser==1.2.2
docker==7.1.0
exceptiongroup==1.3.1
fakeredis==2.32.1
fastapi==0.123.4
fsspec==2025.10.0
graphviz==0.21
greenlet==3.2.4
griffe==1.15.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.14.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lupa==2.6
Mako==1.3.10
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.3.5
oauthlib==3.3.1
opentelemetry-api==1.38.0
opentelemetry-exporter-prometheus==0.59b0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.4
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
pendulum==3.1.0
pluggy==1.6.0
prefect==3.6.4
prometheus_client==0.23.1
psycopg2-binary==2.9.11
py-key-value-aio==0.3.0
py-key-value-shared==0.3.0
pycparser==2.23
pydantic==2.12.5
pydantic-extra-types==2.10.6
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydocket==0.15.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
readchar==4.2.1
redis==7.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==14.2.0
rpds-py==0.30.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
s3transfer==0.16.0
semver==3.0.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.44
starlette==0.50.0
text-unidecode==1.3
toml==0.10.2
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uv==0.9.14
uvicorn==0.38.0
websockets==15.0.1
zipp==3.23.0
